# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BotTalk API
  description: |
    BotTalk 是一个开源的微信推送服务（类似 Server酱），通过微信 iLink Bot 向用户发送消息推送。

    ## 认证方式

    - **SendKey**：用于推送 API，32 位十六进制字符串，支持路径传参、Query 传参和 Bearer Token 三种方式。
    - **Session Cookie**：用于管理 API，通过登录或扫码注册后自动设置，有效期 30 天。

    ## 响应格式

    推送 API 使用 `code` + `message` 风格；管理 API 使用 `success` 布尔值风格。
  version: 1.0.0
  contact:
    name: BotTalk
    url: https://bot-talk.com
servers:
  - url: https://bot-talk.com
    description: 托管服务
  - url: http://localhost:3000
    description: 本地开发 / 自部署
...
components:
  securitySchemes:
    SendKeyQuery:
      type: apiKey
      in: query
      name: key
      description: 通过 query 参数传递 SendKey
    SendKeyBearer:
      type: http
      scheme: bearer
      description: 通过 Authorization Bearer 头传递 SendKey
    SessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: 登录 session cookie
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
    UnauthorizedError:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized. Please log in."
...
    PushLog:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [success, failed]
        ip:
          type: string
        channel_id:
          type: integer
        channel_name:
          type: string
          description: 通过 JOIN 获取的通道名称
        response:
          type: string
          description: JSON 字符串，包含 iLink API 响应
        created_at:
          type: string
          format: date-time
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          nullable: true
        wechat_openid:
          type: string
          nullable: true
        send_key:
          type: string
        role:
          type: string
          enum: [user, admin]
        is_disabled:
          type: integer
          enum: [0, 1]
        rate_limit:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        channel_count:
          type: integer
        push_count:
          type: integer
    AdminStats:
      type: object
      properties:
        total_users:
          type: integer
        today_pushes:
          type: integer
        active_channels:
          type: integer
        disabled_users:
          type: integer
        admin_count:
          type: integer
paths:
...
  /api/push-logs:
    get:
      tags: [Logs]
      summary: 获取推送日志
      description: 获取当前用户最近 50 条推送日志，包含通道名称。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 推送日志列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PushLog'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

  # ════════════════════════════════════════════════════════════════════
  #  Reminders API
  # ════════════════════════════════════════════════════════════════════
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: 用户列表
      description: 获取所有用户列表，含通道数和推送数统计。需要 admin 角色。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: 无 admin 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/users/{id}:
    patch:
      tags: [Admin]
      summary: 修改用户
      description: 修改用户的角色、禁用状态或速率限制。不能修改自己。需要 admin 角色。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户 ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [user, admin]
                  description: 用户角色
                is_disabled:
                  type: integer
                  enum: [0, 1]
                  description: 是否禁用
                rate_limit:
                  type: integer
                  minimum: 0
                  maximum: 100000
                  description: 速率限制（每小时推送条数）
      responses:
        '200':
          description: 更新结果
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      data:
                        type: object
                        properties:
                          id:
                            type: integer
                          email:
                            type: string
                          role:
                            type: string
                          is_disabled:
                            type: integer
                          rate_limit:
                            type: integer
                  - $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: 无 admin 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Admin]
      summary: 删除用户
      description: 删除用户及其所有关联数据（通道、推送日志、提醒等）。不能删除自己。需要 admin 角色。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 用户 ID
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      success:
                        type: boolean
                        example: true
                      data:
                        type: object
                        properties:
                          deleted:
                            type: integer
                            description: 被删除的用户 ID
                  - $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: 无 admin 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/stats:
    get:
      tags: [Admin]
      summary: 仪表盘统计
      description: 获取系统概览统计数据。需要 admin 角色。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 统计数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AdminStats'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: 无 admin 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ════════════════════════════════════════════════════════════════════
  #  Config API
  # ════════════════════════════════════════════════════════════════════
...